Refactor index
