+++ types/chat.ts
@@ -6,3 +6,10 @@ export type Message = {
   isStreaming?: boolean;
   sourceDocs?: Document[];
 };
+
+export type MessageState = {
+  messages: Message[];
+  pending?: string;
+  history: [string, string][];
+  pendingSourceDocs?: Document[];
+};
